refactor: Relocate mailbox route
